Implement some powerup modes
Three powerup modes have been implemented and tested. The three modes are: clearing out a whole row when someone gets a hit, clearing out a whole column when someone gets a hit, and removing 4 other squares when someone gets hit. A player is able to use one of each powerup per game. Once a certain powerup is used and the player gets a hit, that powerup will be disabled for the rest of the game.